--- ghost/core/core/shared/custom-theme-settings-cache.js
@@ -1,3 +1,0 @@
-const {Cache: CustomThemeSettingsCache} = require('@tryghost/custom-theme-settings-service');
-
-module.exports = new CustomThemeSettingsCache();
