+++ ghost/core/core/shared/custom-theme-settings-cache/index.js
@@ -1,0 +1,3 @@
+const CustomThemeSettingsCache = require('./CustomThemeSettingsCache');
+
+module.exports = new CustomThemeSettingsCache();
