+++ ghost/core/package.json
@@ -75,7 +75,6 @@
     "@tryghost/config-url-helpers": "1.0.13",
     "@tryghost/constants": "0.0.0",
     "@tryghost/custom-fonts": "0.0.0",
-    "@tryghost/custom-theme-settings-service": "0.0.0",
     "@tryghost/database-info": "0.3.29",
     "@tryghost/debug": "0.1.34",
     "@tryghost/domain-events": "0.0.0",
