+++ ghost/core/test/unit/shared/custom-theme-settings-cache/service.test.js
@@ -5,7 +5,8 @@ const sinon = require('sinon');
 const {ValidationError} = require('@tryghost/errors');
 const nql = require('@tryghost/nql-lang');
 
-const {Cache, Service} = require('../');
+const Cache = require('../../../../core/shared/custom-theme-settings-cache/CustomThemeSettingsCache');
+const Service = require('../../../../core/shared/custom-theme-settings-cache/CustomThemeSettingsService');
 
 function makeModelInstance(data) {
     const instance = Object.assign({}, data, {
