--- ghost/custom-theme-settings-service/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    Service: require('./lib/CustomThemeSettingsService'),
-    Cache: require('./lib/CustomThemeSettingsCache')
-};
