--- ghost/custom-theme-settings-service/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@tryghost/custom-theme-settings-service",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura --check-coverage -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "@tryghost/nql-lang": "0.6.3",
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/debug": "0.1.34",
-    "@tryghost/errors": "1.3.7",
-    "@tryghost/nql": "0.12.7",
-    "@tryghost/tpl": "0.1.34",
-    "lodash": "4.17.21"
-  }
-}
